+++ webknossos/Changelog.md
@@ -15,6 +15,22 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 ### Breaking Changes
 
 ### Added
+- Added operations for `Project`, `Task` and `TaskType` classes and objects: [#1294](https://github.com/scalableminds/webknossos-libs/pull/1294)
+  - Added operations for `Project` class:
+    - `Project.create()` for creating a new project.
+    - `my_project.delete()` to delete a project object
+  - Added `my_project.update()` to update `priority`, `expected_time` and `is_blacklisted_from_report` values of an existing project object.
+  - Added operations for `TaskType` class:
+    - `TaskType.get_list()` to get all `TaskType` objects.
+    - `TaskType.get_by_id()` to get a `TaskType` by id.
+    - `TaskType.get_by_name()` to get a `TaskType` by name.
+    - `TaskType.create()` to create a new `TaskType`.
+  - Added operations `my_task_type.delete()` method for `TaskType` objects.
+  - Added `Task.get_list()`to get all `Task` objects.
+  - Added operations for `Task` objects:
+    - `my_task.update()` to update a `Task` object.
+    - `my_task.delete()` to delete a `Task` object.
+- Added `TaskExperience` class to store domain and value pair.
 - Added a feature to track attached files of a segmentation layer. Previously, these files were only auto-detected by WEBKNOSSOS based on the location in special folders. Attachments can be added, e.g. `dataset.get_segmentation_layer("segmentation").attachments.add_mesh('path/to/meshfile.hdf5')`. [#1312](https://github.com/scalableminds/webknossos-libs/pull/1312)
 - Added a `with_attachments` parameter to `Dataset.copy_dataset` to copy attachments. [#1312](https://github.com/scalableminds/webknossos-libs/pull/1312)
 - Added a `get_segmentation_layer` method to `Dataset` to get a correctly-typed segmentation layer by name. [#1312](https://github.com/scalableminds/webknossos-libs/pull/1312)